refactor : 게임 재시작/종료
