+++ src/main/java/baseball/BallGame.java
@@ -1,15 +1,24 @@
 package baseball;
 
 import java.util.List;
+import java.util.Scanner;
 
 public class BallGame {
 
-    private static InputNumber playerInputNumbers;
+    private static final String PLAY_RESTART = "게임을 다시 시작하려면 1, 종료하려면 2를 입력하세요 :  ";
+    private static final Scanner scanner = new Scanner(System.in);
 
-    public void play(BaseballResult baseballResult, List<Integer> computerNumbers, List<Integer> playerInputNumbers) {
+    public void play(BaseballResult baseballResult, List<Integer> computerNumbers) {
+        do {
+            baseballResult.initial();
+            //2. get customer value
+            InputNumber playerInputNumbersInitail = new InputNumber();
+            String result = playerInputNumbersInitail.inputBallNumbers();
+            List<Integer> playerInputNumbers = playerInputNumbersInitail.setBallNumber(result);
 
-        baseballResult = getBallCount(baseballResult, computerNumbers, playerInputNumbers);
-        System.out.println(baseballResult.resultPrint());
+            baseballResult = getBallCount(baseballResult, computerNumbers, playerInputNumbers);
+            System.out.println(baseballResult.resultPrint());
+        } while (!baseballResult.isEndGame());
     }
 
     public BaseballResult getBallCount(BaseballResult baseballResult, List<Integer> dealerNumbers, List<Integer> playerInputNumbers) {
@@ -29,5 +38,13 @@ public class BallGame {
             return;
         }
         baseballResult.isBall();
+    }
+
+    public static boolean restart() {
+        System.out.println(PLAY_RESTART);
+        String result = scanner.next();
+        if (result.equals("1")) return true;
+        if (result.equals("2")) return false;
+        return false;
     }
 }
